requirements.txt
PyYAML~=5.4.1
Markdown~=3.3.6
requests~=2.26.0
beautifulsoup4~=4.9.3